move debug feature with default impl to xtext
